Change hash type from uint64 to int64 in FileRecord and update related database queries; adjust hash handling in rebalance logic for consistency

+++ sha2_N-go/rebalance/rebalance.go
@@ -26,7 +26,7 @@ func BalanceAddition(newNode *node.Node, rightNode *node.Node, n int) {
 			continue
 		}
 
-		recordSlot := hash.GetSlot(record.Hash,n)
+		recordSlot := hash.GetSlot(uint64(record.Hash),n)
 		if recordSlot <= newNode.Slot{
 			recordsToMove = append(recordsToMove, record)
 		}
